+++ app/Controllers/subclass-api.controller.php
@@ -1,18 +1,18 @@
 <?php
 require_once './app/Models/subclass.model.php';
 require_once './app/Views/api.view.php';
+require_once './app/Helpers/auth-api.helper.php';
 
 class SubclassApiController {
     private $model;
     private $view;
-
+    private $authHelper;
     private $data;
 
     public function __construct() {
         $this->model = new SubclassModel();
         $this->view = new ApiView();
-        
-        // lee el body del request
+        $this->authHelper = new AuthApiHelper();
         $this->data = file_get_contents("php://input");
     }
 
@@ -24,47 +24,52 @@ class SubclassApiController {
         $arrayClass = ["id_subclass", "name", "author", "id_class"];
         $quant = $this->model->getQuantRegisters();
 
-        if(isset($_GET["filter"])&&!empty($_GET["filter"])&&
+        if(isset($_GET['filter'])&&!empty($_GET['filter'])&&
         isset($_GET['value'])){
-            if(in_array($_GET["filter"], $arrayClass)){
-                $column = $_GET['filter'];
+            if(in_array($_GET['filter'], $arrayClass)){
+                $column = $arrayClass[array_search($_GET['filter'], $arrayClass)];
                 $value = $_GET['value'];
             }else{
-                $column= 1;
-                $value = 1;
+                $this->view->response("Resource not found", 404);
+                die();
             }
         }else{
             $column= 1;
             $value = 1;
         }
 
-        if(isset($_GET["orderBy"])&&!empty($_GET["orderBy"])){
-            if(in_array($_GET["orderBy"], $arrayClass)){
-                $orderBy = $_GET["orderBy"];
+        if(isset($_GET['orderBy'])&&!empty($_GET['orderBy'])){
+            if(in_array($_GET['orderBy'], $arrayClass)){
+                $orderBy = $arrayClass[array_search($_GET['orderBy'], $arrayClass)];
+            }
+            else{
+                $this->view->response("Resource not found", 404);
+                die();
             }
         }else{
             $orderBy=$arrayClass[0];
         }
 
         if((isset($_GET['page']))&&(isset($_GET['limit'])&&!empty($_GET['limit']))){
-            $page = $_GET['page'];
-            $limit = $_GET['limit'];
-            $offset = $page*$limit;
+            if(is_numeric($_GET['page'])&&is_numeric($_GET['limit'])){
+                $page = $_GET['page'];
+                $limit = $_GET['limit'];
+                $offset = $page*$limit;
+            }else{
+                $this->view->response("Resource not found", 404);
+                die();
+            }
         }else{
             $offset = 0;
             $limit = $quant;
         }
 
-        if(isset($_GET["orderBy"])&&!empty($_GET["orderBy"])){
-            if(in_array($_GET["orderBy"], $arrayClass)){
-                $orderBy = $_GET["orderBy"];
-            }
-        }else{
-            $orderBy=$arrayClass[0];
-        }
         if(isset($_GET['cond'])&&!empty($_GET['cond'])){
             if($_GET['cond']==="desc"||$_GET['cond']==="asc"){
-                $cond = $_GET['cond'];
+                $cond = $arrayClass[array_search($_GET['cond'], $arrayClass)];
+            }else{
+                $this->view->response("Resource not found", 404);
+                die();
             }
         }else{
             $cond="asc";
@@ -94,6 +99,11 @@ class SubclassApiController {
     public function deleteSubclass($params = null) {
         $id = $params[':ID'];
 
+        if(!$this->authHelper->isLoggedIn()){
+            $this->view->response("No estas logeado", 401);
+            return;
+        }
+
         $subclass = $this->model->get($id);
         if ($subclass) {
             $this->model->delete($id);
@@ -104,6 +114,11 @@ class SubclassApiController {
 
     public function insertSubclass($params = null) {
         $subclass = $this->getData();
+
+        if(!$this->authHelper->isLoggedIn()){
+            $this->view->response("No estas logeado", 401);
+            return;
+        }
 
         if (empty($subclass->name) || empty($subclass->author) || empty($subclass->id_class)) {
             $this->view->response("Complete los datos", 400);
@@ -116,6 +131,12 @@ class SubclassApiController {
 
     public function editSubclass($params = null) {
         $id = $params[':ID'];
+
+        if(!$this->authHelper->isLoggedIn()){
+            $this->view->response("No estas logeado", 401);
+            return;
+        }
+
         $subclass = $this->model->get($id);
         if ($subclass) {
             $newsubclass = $this->getData();
